requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.9.0
asynclog==0.1.7
asyncpg==0.30.0
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
fastapi==0.115.12
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.10
iniconfig==2.1.0
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
pycparser==2.22
pydantic==2.11.1
pydantic-settings==2.6.1
pydantic_core==2.33.0
PyJWT==2.10.1
pytest==8.3.5
pytest-asyncio==0.24.0
python-dotenv==1.1.0
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlalchemy_utils==0.41.2
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.0
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0.1
